Massive refactor to split up model and dataset, plus fixed tests for dataset and dimensions (most of the remaining system will be broken).
