+++ openspending/model/dimension.py
@@ -11,9 +11,9 @@ class Dimension(object):
     """ A base class for dimensions. A dimension is any property of an entry
     that can serve to describe it beyond its purely numeric ``Measure``.  """
 
-    def __init__(self, dataset, name, data):
+    def __init__(self, model, name, data):
         self._data = data
-        self.dataset = dataset
+        self.model = model
         self.name = name
         self.key = data.get('key', False)
         self.label = data.get('label', name)
@@ -66,9 +66,9 @@ class AttributeDimension(Dimension, Attribute):
     table.
     """
 
-    def __init__(self, dataset, name, data):
-        Attribute.__init__(self, dataset, name, data)
-        Dimension.__init__(self, dataset, name, data)
+    def __init__(self, model, name, data):
+        Attribute.__init__(self, model, name, data)
+        Dimension.__init__(self, model, name, data)
 
     def __repr__(self):
         return "<AttributeDimension(%s)>" % self.name
@@ -78,7 +78,7 @@ class AttributeDimension(Dimension, Attribute):
         distinct values) matching the filter in ``conditions``. """
         query = select([self.column_alias], conditions,
                        limit=limit, offset=offset, distinct=True)
-        rp = self.dataset.bind.execute(query)
+        rp = self.model.bind.execute(query)
         while True:
             row = rp.fetchone()
             if row is None:
@@ -86,12 +86,12 @@ class AttributeDimension(Dimension, Attribute):
             yield row[0]
 
     def num_entries(self, conditions="1=1"):
-        """ Return the count of entries on the dataset fact table having the
+        """ Return the count of entries on the model fact table having the
         dimension set to a value matching the filter given by ``conditions``.
         """
         query = select([func.count(func.distinct(self.column_alias))],
                        conditions)
-        rp = self.dataset.bind.execute(query)
+        rp = self.model.bind.execute(query)
         return rp.fetchone()[0]
 
 
@@ -102,8 +102,8 @@ class Measure(Attribute):
     financial unit, i.e. the amount associated with the transaction or
     a specific portion thereof (i.e. co-financed amounts). """
 
-    def __init__(self, dataset, name, data):
-        Attribute.__init__(self, dataset, name, data)
+    def __init__(self, model, name, data):
+        Attribute.__init__(self, model, name, data)
         self.label = data.get('label', name)
 
     def __getitem__(self, name):
@@ -124,8 +124,8 @@ class CompoundDimension(Dimension, TableHandler):
     have sub-dimensions (i.e. snowflake schema).
     """
 
-    def __init__(self, dataset, name, data):
-        Dimension.__init__(self, dataset, name, data)
+    def __init__(self, model, name, data):
+        Dimension.__init__(self, model, name, data)
         self.taxonomy = data.get('taxonomy', name)
 
         self.attributes = []
@@ -155,7 +155,7 @@ class CompoundDimension(Dimension, TableHandler):
     @property
     def column_alias(self):
         """ This an aliased pointer to the FK column on the fact table. """
-        return self.dataset.alias.c[self.column.name]
+        return self.model.alias.c[self.column.name]
 
     @property
     def selectable(self):
@@ -171,7 +171,7 @@ class CompoundDimension(Dimension, TableHandler):
         column = Column(self.name + '_id', Integer, index=True)
         fact_table.append_column(column)
         if make_table is True:
-            self._init_table(meta, self.dataset.name, self.name)
+            self._init_table(meta, self.model.dataset.name, self.name)
             for attr in self.attributes:
                 attr.column = attr.init(meta, self.table)
             alias_name = self.name.replace('_', ALIAS_PLACEHOLDER)
@@ -210,7 +210,7 @@ class CompoundDimension(Dimension, TableHandler):
         query = select([self.alias], conditions,
                        limit=limit, offset=offset,
                        distinct=True)
-        rp = self.dataset.bind.execute(query)
+        rp = self.model.bind.execute(query)
         while True:
             row = rp.fetchone()
             if row is None:
@@ -220,17 +220,17 @@ class CompoundDimension(Dimension, TableHandler):
             yield member
 
     def num_entries(self, conditions="1=1"):
-        """ Return the count of entries on the dataset fact table having the
+        """ Return the count of entries on the model fact table having the
         dimension set to a value matching the filter given by ``conditions``.
         """
-        joins = self.join(self.dataset.alias)
+        joins = self.join(self.model.alias)
         query = select([func.count(func.distinct(self.column_alias))],
                        conditions, joins)
-        rp = self.dataset.bind.execute(query)
+        rp = self.model.bind.execute(query)
         return rp.fetchone()[0]
 
     def __len__(self):
-        rp = self.dataset.bind.execute(self.alias.count())
+        rp = self.model.bind.execute(self.alias.count())
         return rp.fetchone()[0]
 
     def __repr__(self):
@@ -256,8 +256,8 @@ class DateDimension(CompoundDimension):
         'yearmonth': {'datatype': 'string'},
     }
 
-    def __init__(self, dataset, name, data):
-        Dimension.__init__(self, dataset, name, data)
+    def __init__(self, model, name, data):
+        Dimension.__init__(self, model, name, data)
         self.taxonomy = name
 
         self.attributes = []
